Move graphql schema
